Desktop/New folder (4)/twitter_bot/requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
schedule==1.2.0
